required.txt added

+++ store/models.py
@@ -83,8 +83,8 @@ class CartItem(models.Model):
 class Coupons(models.Model):
     coupon_code = models.CharField(max_length=10)
     expired = models.BooleanField(default=False)
-    discount_price = models.FloatField(default=100)
-    minimum_amount = models.FloatField(default=500)
+    discount_price = models.PositiveIntegerField(default=100)
+    minimum_amount = models.PositiveIntegerField(default=500)
     expiry_date = models.DateField(null=True,blank=True)
 
     def __str__(self):
